+++ boostPython/cloudComPy.cpp
@@ -721,6 +721,8 @@ BOOST_PYTHON_MODULE(_cloudComPy)
 
     def("isPluginPCV", &pyccPlugins::isPluginPCV, cloudComPy_isPluginPCV_doc);
 
+    def("isPluginCSF", &pyccPlugins::isPluginCSF, cloudComPy_isPluginCSF_doc);
+
     def("isPluginRANSAC_SD", &pyccPlugins::isPluginRANSAC_SD, cloudComPy_isPluginRANSAC_SD_doc);
 
     def("computeCurvature", computeCurvature, cloudComPy_computeCurvature_doc);